refactor: split PlayProgressWidget into PlayButton, ProgressBar, ActionButtons
